# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App",
        "version": "0.1"
    },
    "paths": {
...
        "/api/app/cart": {
            "get": {
                "tags": [
                    "APP Cart"
                ],
                "summary": "Lấy giỏ hàng",
                "operationId": "4347ff5a19d81b7e5f81697065fac31a",
                "responses": {
                    "200": {
                        "description": "Danh sách sản phẩm trong giỏ hàng"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/cart/store": {
            "post": {
                "tags": [
                    "APP Cart"
                ],
                "summary": "Thêm sản phẩm vào giỏ hàng",
                "operationId": "8fd36d3ad1133dcd2de9467fbe3cf9bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sản phẩm đã được thêm vào giỏ hàng"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/cart/update": {
            "put": {
                "tags": [
                    "APP Cart"
                ],
                "summary": "Cập nhật giỏ hàng",
                "operationId": "b9d1747da76a4320f6ada4b1a79753ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật số lượng thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/cart/delete/{id}": {
            "delete": {
                "tags": [
                    "APP Cart"
                ],
                "summary": "Xóa sản phẩm khỏi giỏ hàng",
                "description": "Xóa một sản phẩm khỏi giỏ hàng của người dùng đã đăng nhập.",
                "operationId": "b046208a03c4ab82c7f204c6d9c1be2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của sản phẩm cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa sản phẩm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa sản phẩm thành công."
                                        },
                                        "deleted_count": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/cart/clear": {
            "delete": {
                "tags": [
                    "APP Cart"
                ],
                "summary": "Dọn sạch giỏ hàng",
                "operationId": "829aab107b3ecaa84092e5dd18cfed5b",
                "responses": {
                    "200": {
                        "description": "Giỏ hàng đã được dọn sạch"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/cart/checkout": {
            "post": {
                "tags": [
                    "APP Cart"
                ],
                "summary": "Đặt hàng từ giỏ hàng",
                "description": "Người dùng có thể đặt hàng các sản phẩm từ giỏ hàng của mình.",
                "operationId": "0c46ab596de32fb683649e8c1ce88236",
                "requestBody": {
                    "description": "Danh sách ID sản phẩm cần đặt hàng",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_ids": {
                                        "description": "Mảng chứa ID của các sản phẩm cần đặt hàng",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đặt hàng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Đặt hàng thành công"
                                        },
                                        "order": {
                                            "description": "Thông tin đơn hàng",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu không hợp lệ."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Có lỗi xảy ra, vui lòng thử lại sau."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}